Controller 구분, 홈페이지, 관리자 페이지, 사용자 전환 신청 구현 완료
