+++ OneclickDonation/src/main/java/com/example/OneclickDonation/admin/AdminController.java
@@ -1,11 +1,10 @@
 package com.example.OneclickDonation.admin;
 
 import com.example.OneclickDonation.admin.dto.UpgradeAdminDto;
-import com.example.OneclickDonation.member.dto.MemberDto;
 import com.example.OneclickDonation.post.dto.PostDto;
 import com.example.OneclickDonation.post.service.PostService;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -16,13 +15,17 @@ import org.springframework.web.bind.annotation.*;
 
 @Slf4j
 @Controller
-@RequestMapping("admin")
-@RequiredArgsConstructor
 public class AdminController {
     private final AdminService service;
     private final PostService postService;
 
-    @GetMapping
+    @Autowired
+    public AdminController(AdminService service, PostService postService) {
+        this.service = service;
+        this.postService = postService;
+    }
+
+    @GetMapping("/admin")
     public String adminPage(
             Pageable pageable,
             Model model
@@ -32,7 +35,7 @@ public class AdminController {
         return "admin/home";
     }
 
-    @GetMapping("/upgrades")
+    @GetMapping("/admin/upgrades")
     public String upgradeRequests(
             Pageable pageable,
             Model model
@@ -40,25 +43,5 @@ public class AdminController {
         Page<UpgradeAdminDto> upgradeList = service.listRequests(pageable);
         model.addAttribute("upgradeList", upgradeList);
         return "admin/upgrade";
-    }
-
-    // 수락
-    @PostMapping("/upgrades/{id}/accept")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public UpgradeAdminDto approve(
-            @PathVariable("id")
-            Long id
-    ) {
-        return service.acceptUpgrade(id);
-    }
-
-    // 거절
-    @DeleteMapping("/upgrades/{id}/accept")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public UpgradeAdminDto disapprove(
-            @PathVariable("id")
-            Long id
-    ) {
-        return service.rejectUpgrade(id);
     }
 }
